+++ manage/docker-tiny/setup.sh
@@ -8,7 +8,7 @@ LOCAL_DIR=".local/docker-tiny"
 LOG_DIR="$LOCAL_DIR/log"
 CONFIG_FILE="$LOCAL_DIR/local.py"
 CONFIG_TEMPLATE_FILE="setup/local.docker.dev.py"
-DB_FILE="$LOCAL_DIR/db.sqlite3"
+DB_FILE="$LOCAL_DIR/db/db.sqlite3"
 DC_FILE="setup/docker-compose.tiny.yml"
 DC="docker-compose -f $DC_FILE"
 
@@ -24,7 +24,10 @@ fi
 # Create DB file so Docker doesn't make it a directory
 if [[ ! -e $DB_FILE ]]; then
     echo "Creating DB file ..."
+    mkdir -p "$(dirname "$DB_FILE")"
+    chmod 777 "$(dirname "$DB_FILE")"
     touch "$DB_FILE"
+    chmod 666 "$DB_FILE"
 fi
 
 # Add version file